vendor: upgrade storybook
