--- .storybook/presets.js
@@ -1,1 +1,0 @@
-module.exports = ['@storybook/addon-docs/preset'];
