+++ package.json
@@ -18,60 +18,60 @@
     "storybook": "start-storybook -p 6006"
   },
   "devDependencies": {
-    "@babel/cli": "^7.7.7",
-    "@babel/core": "^7.7.7",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.7.4",
-    "@babel/plugin-proposal-optional-chaining": "^7.7.5",
-    "@babel/preset-env": "^7.7.7",
-    "@babel/preset-react": "^7.7.4",
-    "@babel/preset-typescript": "^7.7.7",
-    "@borealisgroup/eslint-config-ts": "1.1.3",
-    "@material-ui/core": "^4.8.1",
-    "@material-ui/icons": "^4.5.1",
-    "@storybook/addon-docs": "^5.3.0-rc.0",
-    "@storybook/addon-knobs": "^5.3.0-rc.0",
-    "@storybook/addons": "^5.3.0-rc.0",
-    "@storybook/cli": "^5.3.0-rc.0",
-    "@storybook/react": "^5.3.0-rc.0",
-    "@storybook/theming": "^5.2.8",
-    "@types/faker": "^4.1.8",
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.8.4",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
+    "@babel/plugin-proposal-optional-chaining": "^7.8.3",
+    "@babel/preset-env": "^7.8.4",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@borealisgroup/eslint-config-ts": "1.1.4",
+    "@material-ui/core": "^4.9.1",
+    "@material-ui/icons": "^4.9.1",
+    "@storybook/addon-docs": "^5.3.10",
+    "@storybook/addon-knobs": "^5.3.10",
+    "@storybook/addons": "^5.3.10",
+    "@storybook/cli": "^5.3.10",
+    "@storybook/react": "^5.3.10",
+    "@storybook/theming": "^5.3.10",
+    "@types/faker": "^4.1.9",
     "@types/is-hotkey": "^0.1.1",
-    "@types/jest": "^24.0.24",
+    "@types/jest": "^25.1.1",
     "@types/lodash": "^4.14.149",
-    "@types/node": "13.1.1",
+    "@types/node": "13.7.0",
     "@types/prismjs": "^1.16.0",
-    "@types/react": "16.9.17",
-    "@types/react-dom": "16.9.4",
-    "@types/styled-components": "^4.4.1",
+    "@types/react": "16.9.19",
+    "@types/react-dom": "16.9.5",
+    "@types/styled-components": "^4.4.2",
     "awesome-typescript-loader": "^5.2.1",
     "babel-loader": "^8.0.6",
     "babel-plugin-add-react-displayname": "^0.0.5",
     "babel-plugin-inline-json-import": "^0.3.2",
-    "babel-plugin-styled-components": "^1.10.6",
+    "babel-plugin-styled-components": "^1.10.7",
     "eslint": "^6.8.0",
-    "eslint-plugin-mdx": "^1.6.5",
+    "eslint-plugin-mdx": "^1.6.8",
     "faker": "^4.1.0",
-    "jest": "^24.9.0",
-    "lerna": "^3.19.0",
+    "jest": "^25.1.0",
+    "lerna": "^3.20.2",
     "react": "^16.12.0",
     "react-docgen-typescript-loader": "^3.6.0",
     "react-dom": "^16.12.0",
-    "rollup": "^1.27.14",
+    "rollup": "^1.31.0",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-node-builtins": "^2.1.2",
     "rollup-plugin-node-globals": "^1.4.0",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-terser": "^5.1.3",
+    "rollup-plugin-terser": "^5.2.0",
     "rollup-plugin-typescript2": "^0.25.3",
     "slate": "0.57.1",
     "slate-history": "0.57.1",
     "slate-hyperscript": "0.57.1",
     "slate-react": "0.57.1",
-    "styled-components": "^4.4.1",
-    "ts-jest": "^24.2.0",
+    "styled-components": "^5.0.0",
+    "ts-jest": "^25.1.0",
     "ts-loader": "^6.2.1",
-    "typescript": "3.7.4"
+    "typescript": "3.7.5"
   }
 }
